Extracted WooPosItemsEmptySearchQueryState component for handling empty search queries
